Fix circular dependency issue
